==> Book/07-conclusiones.Rmd <==
# Conclusiones

Como conclusión a este proyecto, podemos recalcar que el análisis de datos es un área enorme de la informática. Hemos extraído información de numerosas maneras, a pesar de que el dataset no es tan extenso como las compañias suelen tratar mediante Big Data.

Hemos aprendido a construir un dataset desde cero, y hemos aplicado preprocessing y filtrado en cada nueva herramienta de análisis de datos que hemos tenido que usar. Esta fase del análisis es la más importante, y es la que menos suelen pensar al iniciarse en esta disciplina.

Una de las técnicas más relevantes y la que nos ha ofrecido mejores resultados es el análisis de conceptos formales. Incluso ha servido para poder extraer reglas fácilmente y poder ser tratadas como reglas de asociación, con implicaciones mucho mejor desarrolladas y significativas.

En definitiva, trabajar con este proyecto nos ha servido para involucrarnos en una tarea real de análisis de datos, y esperemos que no sea la última, porque esta disciplina es una de nuestras favoritas.
